Move APICall enum to a separate class
